requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.38.0
opencv-python>=4.9.0
scenedetect>=0.6.3
Pillow>=10.2.0
numpy>=1.26.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0
tqdm>=4.66.0
pytest>=8.0.0
